lua: fix intractable memory management for zmsgs
zmsg-lua module tried to be clever about saving the original
zmsg by pretending we freed the message but actually saving
the original in a zmsg_info object.

This turned out to be too difficult to get right, and "something"
at a higher level seemed to be freeing or modifying the zmsgs
such that the subsequent zmsg_destroy() invoked by Lua garbage
collection was getting memory errors. (We got away with these
most of the time, but only by luck)

This patch chickens out and saves a copy of each zmsg in the
zmsg_info structure.