wreck: Use configured path to wrexecd if it exists.
Instead of only relying on #define version of WREXECD_PATH, track
the kvs value 'config.wrexec.wrexecd_path' and use that as the
path to wrexecd if it exists. If the path does not exist, the
fallback to the default (system) location.

The wrexec comms module now installs a watch on this value, so
the path to wrexecd can be set at runtime.