wreck: fix bug in use of posix module in lua plugin
In 01-env.lua wreck/lua plugin, the posix module was being
loaded but assigned to the variable `p' instead of `posix'.
The code itself referenced the `posix' table, but this only
worked becuase Lua 5.1 effect of instantiating a global
'posix' table when loading the lua-posix module.

In 5.2, modules can no longer instantiate global tables, so
the 01-env.lua plugin was broken.

This patch simply assigns the lua-posix module table correctly
to the 'posix' variable. (now a local)