broker: make event socket connection synchronous
When events are being distributed by ipc:// or tcp:// endpoint,
block until the socket is connected.  Events may still be missed
by the node, but this guarantees that all nodes are connected to
the event network before the 'live.ready' event is generated and
the command/shell is started.

Revert the 250ms delay that was working around spuriously failing
tests.

Fixes issue #38