broker: handle events sent to ring peers
A consequence of Cc'ing events to newly connected peers is that
ring network peers connect to the same request socket as TBON peers.
Thus an additional redundant source of events is possible.

Allow this but instead of shutting off the parent when the first
duplicate is received, set a flag when the first event is received
on the primary event socket, then when events are received on other
sockets and this flag is set, send that peer (on whatever socket)
a mute request.