broker: allow rank-addressed request loopback
If a request is address to a specific rank, and the rank matches
that of the sender, and the target is the same service that sent
the message, allow the message to loop back to the sender.

This is a useful way to simplify some message flows in modules,
for example if a module is both producing and consuming particular
message.