broker: destroy inproc://dq-req on module unload
We were waiting around for these sockets to be cleaned up when the
context was destroyed, but now that modules are dynamic, that's a
resource leak.

In addition, while testing modctl, occasionally we would see a
fatal bind failure:

  cmbd: zsocket_bind: inproc://dq-resp-0x10059c0: Address already in use

Possibly the "unique" value based on plugin pointer value was
getting reused.  In any case, closing sockets seems to have fixed it.