libflux: implement 4-byte matchtag per RRC 3
RFC 3 has been updated for a 32-bit matchtag.  Update the the
protocol code, handle code, and libkvs.c to match.

Use FLUX_MATCHTAG_NONE instead of hardwiring a matchtag of 0 when
matching is not used.

Modify the "tag pool" to allocate from 256 blocks of 2^24 tags.
So the pool has the same capacity as before, but code that will
manage a block of concurrent requests can allocate a large block
at one time.