libflux: add flux_json_multrpc(), use in flux-module
flux_json_multrpc() is a variant of flux_json_rpc() that takes
a nodeset instead of a nodeid.  It allocates a block of
matchtags the size of the nodeset, then keeps 'fanout'
requests to nodeset members active until all nodes have been
queried.  A user-supplied callback is made for each response,
if configured.

An error from any of the RPC's causes the whole function to fail,
though it will continue until all transactions are complete and
return -1 with errno set to the highest errnum value received.

Replace the nodeset iteration code in flux-module with calls
to flux_json_multrpc().