cmbd: fix crash on invalid ping packet
An invalid ping packet, i.e. one that did not consist of a true
json object, would cause the cmbd to crash when the "route" was
attempted to be inserted into th response. The crash occurs,
for instance if a json array or other non-object was sent as the
ping payload.

To fix, we check within the ping_req_cb() for a json object type
of 'json_object' and return a protocol error if not.