libflux: rework flux_putmsg_list()
Rename flux_putmsg_nomatch() to flux_putmsg_list().
Change the list parameter from a (zlist_t **) to a (zlist_t *)
and don't destroy the list internally.

In flux_recvmsg_match(), change the list parameter from
a (zlist_t **) to a (zlist_t *) and presume that if the parameter is
non-NULL it points to a valid list.