kvs: add kvs_unwatch()
Add kvs.unwatch handler to cancel a watcher in the kvs module.

Rework wait_destroy_byid(), renaming it to wait_destroy_match().
Now it finds matching messages by looking at the zmsg stored in the
wait_t and calls a user-supplied match function.

Fix the "kvs.disconnect" handler so it calls wait_destroy_match()
to with a match function that matches all requests from the sender.

The KVS unwatch handler calls wait_destroy_match() with a match function
that matches on topic string, sender, and KVS key.

The unwatch client code makes the unwatch RPC and also deregisters
the watch handler, if any.