handle: use veb for matchtag pool
Replace the pool of 256 matchtags with a 64K pool.
Using the veb data structure, this uses about 9K bytes.
Searching it for a free tag is O(log m) where m is 64k.

If a block of consecutive matchtags is required, fall back to the
old allocator where 255 blocks of 2^24 are available, no partial
block support.