util: fix subprocess duplicate output on exec failure
When a process using the subprocess_manager code is running
with fully buffered output, a child which fails during exec(2)
may flush the parent's buffers on the subsequent call to exit(127).

This is only a problem when child shares parent stdio streams,
and will be fixed when the subprocess code properly redirects
child fds to pipes or other destinations. For now, we close
stdout and stdin before exit() to avoid this issue.