testsuite: add some multi-rank kvs tests
Dynamically size the KVS test session so that it is one more than
the number of processors reported in /proc/cpuinfo (minimum of 4).

Add some tests for put and unlink to verify that all ranks immediately
see the new key/don't see the unlinked key.  Use version synchronziation
so that eventually consistent semantics are accounted for.

Change the commit/fence test to run 8 threads per rank instead
of 100 all on rank 0.  For the fence test, supply the total task
count as an argument.