kvs: fix assertion c->state == COMMIT_STORE
Since a fence operation is terminated by an event, and events are
delivered in a non-deterministic order, a client may exit a fence
operation and enter antother fence/commit before the internal state
the previous fence is cleared from upstream ranks.

The short term fix is to remove the old state and proceed rather than
assert.  Longer term we may want to decouple the fence/commit code and
protocol so that cleanup and hashing of state is less fragile.

Fixes issue #109