libflux: rework reactor back-end interfaces
Add new ops in the reactor backend to support registration of handle
message callback on demand.  Before this change, the reactor
implementations always had at least the handle callbacks registered,
so we had to explicitly terminate the reactor when there were no more
handlers registered in at the generic reactor level.

Now we don't need to know what handlers are registered for zsock,
fd, and timer and can have the underlying implementations manage those,
which they already do anyway.

The primary function of the generic handler code is for registration
and dispatch of message handlers, as well as mapping the generic
operations to implementation specific ones (but directly now).