handle: fix issue #146 tkvsunwatch test failure
Add a "purge" operation that allows messages to be purged from
the pugmsg queue if they match the supplied flux_match_t.
Then when a matchtag is retired, purge any matching responses
that were queued out of order on the putmsg queue.

This handles the kvs_unwatch() case described in #146.