libutil: add initial coproc class
This class is a wrapper around getcontext/makecontext/swapcontext.
It hides the gory argument passing details.

Coprocesses are created with a fixed 16K stack size.
The stack has a small patterned redzone at beginning and end that is
checked when coprocesses return or yielded.

Coprocess sigmask is inherited from the parent.

A unit test driver verifies basic operation.