test: tighten up tkvswatch mt test
The tkvswatch mt test creates some number of threads watching the
same key.  The main then moves that key through a sequence of values.
The watch callbacks terminate their thread's reactor when the final
value is received, and the thread exits.  The main joins with all
the threads at the end.  The test would then hang in pthread_join
if any watch callback did not receive the last value.

It should be noted that commits can be coalesced on the KVS master
for performance, and since kvs_watch callbacks are driven
by commits, the value may "jump" from the callback's point of view.
The invariant being tested is that the slave cache is eventually
consistent; that is the last-put value will always be received.

To aid in tracking down issue #151, we now also check that the main
watch callback is not called when the key hasn't changed.  We also
now watch two stable keys (one with value, one without) and ensure
that while the main key is changing, we don't receive callbacks for
those.