reactor: restructure back end message handler
In flux_t handle reactor implementations, call the message handler
when a message can be read, and then let the handler read it,
as opposed to reading it in the implementation and passing the message
to the handler.

The new approach is more in line with how the other handlers work,
and consistent with how the reactor pattern is typically implemented.
In addition, some simplification of reactor implementations results.