test: fix t0002-request.t coproc test
The sycnhronization was wrong in this test.

Replace a usleep() with a new op "req.count" that allows the test
driver to poll to see when the other driver thread has gotten itself
stuck.  Only then can we proceed with
(a) the test to make sure the 'coproc' module remains responsive
to other ops, and
(b) sending the message to unblock the stuck RPC and verify that
it resumes properly.

Fixes #163