module: replace special EOF with 'shutdown' method
Fix the problem described in issue #164 where 'flux module remove'
would cause the module's flux_reactor_start() to return an error.

The handshake for removing a module was to send "EOF" (a zero size
message) on the module' request socket.  Upon receiving this, the
module's handle implementation would stop the reactor and send
an EOF back.  The broker would then call pthread_join(), expecting
the module to have exited its main function.  (This is a rather
hopeful supposition).

The internal reactor implementation had been improved recently so a
message handler would call flux_recvmsg() directly rather than be
passed a message that had been read internally.  Handling the EOF
internally stopped working properly, with flux_recvmsg() returning
NULL and this going down various error paths in the module that
stopped the reactor with an error.

In the interest of moving towards a better shutdown architecture,
I implemented a default method called "shutdown".  Each module gets
a default handler for it that calls flux_reactor_stop(), but it
is overrideable (thus a module could execute its own, perhaps
distributed, shutdown procedure before allowing itself to be removed).
This change allows the shutdown to be handled properly without any
special casing in the handle implementation.

The broker handler is unchanged - upon comms module main() exit,
an EOF message is returned to the broker, as before.

Fixes #164.