doc: fix inline POSIX 2008 reference in cmbd.c
This citation came from IBM in a slurm discussion.  Upon looking
it up, the "page number" is nonexistent, and when looked up as a line
number (the lines are numbered in the pdf), it referred the exit
code for the shell 'export' command.  Substitute a section reference
that talks about command exit status.

http://cfajohnson.com/pdf/Portable_Operating_System_Interface-POSIX.pdf