broker: eliminate cmb.pub and send naked events
Now that we have a decent message protocol, we can send events
without further encapsulation on the TBON, routing them to rank 0
for publishing.

The cmb.pub RPC goes away.  Now there is no acknowledgement coming back
from the rank 0 node indicating that the event was successfully
published, but then rank 0 has no way of knowing that the pub was
successful anyway.  And this is much simpler.