libflux: add FLUX_MSGFLAG_UPSTREAM
Add an extra 'flags' argument to flux_msg_set_nodeid() and
flux_msg_get_nodeid() that can be 0 or FLUX_MSGFLAG_UPSTREAM.
This new flag is vaild only for requests, and indicates that the
nodeid is the sending rank, and that the request should be treated
as though the rank were FLUX_NODEID_ANY except it should be
routed upstream of the sending rank.  Add a few unit tests for this.

Modify broker so that it no longer denies loopback requests, and
instead honors this flag.

Add a reserved nodeid: FLUX_NODEID_UPSTREAM which is only used in
the higher level functions flux_json_request(), flux_json_rpc(), etc..
Internally these functions set the nodeid of the message to the sending
rank, with FLUX_MSGFLAG_UPSTREAM set.

Change all the places where comms modules relied on the old behavior
of FLUX_NODEID_ANY to FLUX_NODEID_UPSTREAM.

Flux RFC 3 will need to be modified to reflect
  FLUX_MSGFLAG_UPSTREAM
  FLUX_NODEID_UPSTREAM (decreases valid nodeid range by one)
  remove section describing the old loopback behavior