test: t0001-basic run flux comms not flux up
With the previous commit, we are starting the initial program
without synchronizing with the live module.  Therefore, tests
that verify that the live module's KVS objects contain the full
nodeset are racy and will fail.

I think the goal here is just to make sure that the broker is responding
so 'flux comms info' is a valid substitute with fewer dependencies,
appropriate for an early test.