broker: drop event socket for modules
Send events to modules over the PAIR socket, which is now shared for
all types of messages.  Implement subscription management in the
broker (cmb.sub / sub.unusub RPC).  The handlers for these RPC's
extract the requestor's uuid and call back into the module code
to manage a list of subscription topics for the module with that
uuid.  Events are then routed to subscribing modules via
module_event_mcast()