handle: add flux_open(), flux_close()
Add new flux_open(), flux_close() functions as described in issue #178.
Handle implementations (connectors) are loaded by name from the connector
search path, e.g. the 'local implementation is expected to be named
'local.so' and to export a connector_init() function.  The search path
is taken from the FLUX_CONNECTOR_PATH environment variable, or if unset,
from the $(fluxconnectordir) install directory.

If the uri argument to flux_open() is NULL, a default value is may be
supplied via the FLUX_URI environment variable.  Otherwise the local
connector is with a NULL path, which local interprets as "use $FLUX_TMPDIR".

The flux.c command driver is modified to
- set FLUX_CONNECTOR_PATH, optionally modified by the --connector-path option
- set FLUX_URI, if --uri option is supplied
- set FLUX_HANDLE_TRACE=1, if --handle-trace option is supplied