libflux: add flux_msg_get_payload_json_str()
Add message accessor functions for json payloads as strings
rather than json-c-specific json objects.  The new functions are

flux_msg_get_payload_json_str()
flux_msg_set_payload_json_str()

In 'get', a payload copy on the critical path is avoided by making
the payload (on the wire) a valid C string.  This is not in conflict
with anything in RFC 3, but the RFC should be updated to state that
that the last characer of a JSON payload MUST be %x00.

Reimplement the functions we wish to deprecate in terms of the above:

flux_msg_get_payload_json()
flux_msg_set_payload_json()

Add unit tests for all of the above.