libflux: flux_msg_get_topic() return static string
Change the prototype of flux_msg_get_topic() so that it returns
a string that is owned by the message rather than allocating a copy.
This is achieved by appending a NULL to the topic string on the wire.
RFC 3 should be updated to reflect this.

This drops a malloc on the critical path and makes the API a little
less prone to memory leaks.