libflux: change flux_event_send, flux_event_recv
Change flux_event_send() and flux_event_recv() prototypes
to handle zmsg_t's encoded/decoded with flux_event_encode()
and flux_event_decode().

Update users of flux_event_send() to
- convert json_object payload (if any) to string with Jtostr (obj)
- create zmsg with flux_event_encode()
- send it with flux_event_send()

Update users of flux_event_recv() to
- recv zmsg with flux_event_recv()
- decode it with flux_event_decode ()
- convert payload (if any) to json_object with Jfromstr (str)

There is one complication with converting to flux_event_recv().
Since flux_event_decode() requires that the caller knows in advance
whether there is a payload, in a few cases, such as in flux-event
and the lua bindings, where the code is not handling a specific
protocol, is was necessary to use the lower level message payload
accessor instead of flux_event_decode().