re-organization for dependencies
libcore is now built in src/common, the header has also moved
libmrpc is now built in src/modules/libmrpc, it can not be built anywhere
else, at all.  It must be built between the kvs and the modctl modules.  How
precisely we avoided this before... all I can think is that the modctl lib
must have been assuming the symbol would be available at runtime, and it
was, so it worked.  Now that it's building shared only, it checks, and if
libmrpc is built in any other order, it fails.

This seems almost ready for merge, pending review, except for one problem:
PROBLEM: flux keygen fails

make check passes all tests, but fails on keygen with either:
    flux-keygen: mkdir
    /g/g12/scogland/projects/flux/build/src/cmd/.libs/../../etc/flux: No such file
    or directory
or:
    flux: `keygen' is not a flux command.  See 'flux --help'

The first happens when an install path is set with configure, the second when
it isn't.