resolved several hidden unresolved symbol errors
Both the kvs and core libraries are now built with flags that require all of
their symbols to resolve at link time for them rather than load time.  This
should probably be a requirement on every shared library we build, if only for
debugging purposes.  The other main change is that the flux command now
detects when it is being invoked by a libtool run script, by checking the path
for containing a .libs component, and adds an extra '../' in that case.  This
was the cause of the keygen error from the previous commit.