libflux-core.so is now stand-alone
Since the KVS symbols are in the module library now, the only symbols left to
be exported by the core were the flux_* symbols.  Since all of those are in
libflux, libflux is now linked into libflux-core but *not* into
libflux-internal.  This way there will only be one copy of the flux_* symbols
in a final application, and they will be exported by the so.