libflux-core.so is now stand-alone
Since the KVS symbols are in the module library now, the only symbols left to
be exported by the core were the flux_* symbols.  Since all of those are in
libflux, libflux is now linked into libflux-core but *not* into
libflux-internal.  This way there will only be one copy of the flux_* symbols
in a final application, and they will be exported by the so.

libflux-internal is static only, travis is verbose

In an attempt to be completely perscriptive about how the linking process is
accomplished, the libflux-internal build is now static only.  I believe the
failure on travis was due to it building internal as an SO, and not a .a, by
default and then failing to pull in the transitive dependencies specified in
the libtool dependency list for the linking of libflux-core.  The core library
also now explicitly references all of the external shared libraries it may
reference, which is probably a good idea anyway.

For debugging issues like this in the future, the travis make command now
includes the V=1 flag to print more complete build information.

dependencies no longer exported by libkvs caused test builds to fail on travis, setting up a reproducing environment to test in

split kz, forkzio from libzio

Move kz to src/modules/libkz and build between src/modules/kvs
(which it requires) and src/modules/wreck (which requires it).
Build as a noinst convenience library for now as kz_*() is not
currently exported as part of the public api.

Move forkzio.[ch] to src/cmd and link directly with flux-zio as that
is its only user.

dependencies and leaking symbols I think are resolved, re-arranging kz helped a great deal, just had to ensure that it didn't include any symbols that should not be leaked.

zio factored out of common, now a fluxlib