libflux-internal is static only, travis is verbose
In an attempt to be completely perscriptive about how the linking process is
accomplished, the libflux-internal build is now static only.  I believe the
failure on travis was due to it building internal as an SO, and not a .a, by
default and then failing to pull in the transitive dependencies specified in
the libtool dependency list for the linking of libflux-core.  The core library
also now explicitly references all of the external shared libraries it may
reference, which is probably a good idea anyway.

For debugging issues like this in the future, the travis make command now
includes the V=1 flag to print more complete build information.