Integrate Job Status and Control (JSC) interface
Provide a high-level abstraction to allow monitoring and controlling
of the status of Flux jobs. Designed to expose the job status and
control abstraction in a way to hide the underlying data layout of
job information stored within Flux's KVS data store.
Expect that schedulers and runtime tools will be its main users. This
abstraction provides the producers of job information including
a task and program execution service module including wreckrun with
an opportunity to change and optimize the data layout of jobs within
the KVS without presenting major impacts to the implementation
of the schedulers and runtime tools.

The main design considerations are the following:
   - Abstract out implementation-dependent job states;
   - Provide flexible and easily extensible mechanisms to pass job info;
   - Use a minimalistic API set.

Build on Job Control Block (JCB):
Our data schema containing the information needed to manage a
particular job. Contains information such as jobid, resources owned
by the job, as well as the processes spawned by the job. JSC converts
the raw information on a job into an JCB, implemented as an JSON
dictionary object. The current JCB structure is described in tables
in README.md.

Provide three main calls of JSC:
   - jsc_notify_status () Register a callback to the asynchronous
        status change notification service. callback will be invoked
        when the state of a job changes;
   - jsc_query_jcb () Query the key attribute of JCB of jobid;
   - jsc_update_jcb () updates the key attribute within JCB of jobid.

Provide standalone utility and testers:
Created an utility command to facilitate testing: flux-jstat. Its
usage is the following:

   Usage: flux-jstat notify
        flux-jstat query jobid <top level JCB attribute>
        flux-jstat update jobid <top level JCB attribute> <JCB JSON>

flux-core/t/t2001-jsc.t also contains various test cases that use
this utility.