libflux: add new json_str based request functions
Add new request functions:
  flux_request_encode()
  flux_request_decode()
  flux_request_send()
  flux_request_sendto()
  flux_request_recv()

These functions are divided into encode/decode and send/recv groups,
as was done with the event functions.  In addition, matchtags and
destination nodeid are treated as a property of send/recv not encode/
decode (this seemed natural even though technically they are part
of the message encoding).

And unit tests for the encode/decode functions.