libflux: update flux_module encode/decode for new API
Also get rid of json_objects in the module API.  As a result, several
functions changed prototypes, and an intermediate object flux_modlist_t
was introduced to simplify the rather convoluted set of functions we
had for manipulating lists of modules.

Update users (broker, flux-module command, unit and system tests).