build: revert earlier libflux test_ldadd changes
Now that libflux no longer depends on libjsonc (which in turn
depends on libflux), stop linking libflux unit tests against
  ../libflux-core.la ../libflux-internal.la
to resolve the circular dependency.

That was confounding an edit/make-check/edit/makecheck workflow
in libflux (one had to remake the ../*.la's in order for
unit tests to see changes).