libflux: add fatal error callback hook
Applications may register a fatal error handler with the flux_t handle.
If an internal error occurs such that using the handle afterwards
would be inadvisable, the user's handler will be called.  If the
handler returns, internal errors will still be propagated back to
callers.  The user's handler will only be called once.

Internal libflux-core code, or the application, may call
flux_fatal_error() to initiate fatal error handling.  If
no handler was registered, or the handler returns, flux_fatal_error()
will return, and error handling proceeds as with non-fatal errors.

A macro FLUX_FATAL(h) is provided which calls flux_fatal_error()
with the calling  __FUNCTION__ and a message of strerror (errno).