broker: exec: initial subprocess io support
Handle subprocess io in the simplest way possible -- by sending output
as replies to the initial "exec" message. Alternate io methods may be
registered in the future.

The broker's zloop gets registered with the subprocess manager, and a
simple io callback sends all io as replies to the initial launch message.
The output json blob is tagged with the rank of the current broker.