libflux: drop pointer from flux_msg_t typedef
Rather than defining the flux_msg_t type as a pointer to an
incomplete struct, define it directly to the incomplete struct.
This allows const to work properly and avoids obscuring the
fact that we are dealing with a pointer.

Add const qualifier to function parameters where appropriate,
especially in message.[ch], request.[ch], response.[ch], event.[ch].

Update users:  broker module handle, and connectors.