libflux/rpc: avoid recycling aborted matchtags
When a flux_rpc_t is destroyed, it is not safe to free its matchtag
block unless the RPC has completed.  If it has not completed, there may
still be responses in flight that could becone intermingled with
another rpc if the matchtags are freed and reused.

To avoid this situation we simply leak the matchtag block when
flux_rpc_destroy() is called if responses have not been received.
There are currently 2^16 single matchtags available per handle for
singleton rpcs, and 2^8 matchtag blocks for multiple rpcs, thus
65535 single RPC's can be prematurely destroyed before the handle cannot
be used for RPC's anymore, or 256 multiple RPC's.  This is not ideal
but it is the simplest alternative at the moment, and converts what
may be obscure distributed failures into a hard EBUSY flux_rpc() failure.

A better solution is to add RPC cancellation to RFC 6.

This came up because t/loop/multrpc.t began failing.  It contains a
subtest where, after sending a block of requests the flux_rpc_t is
destroyed before all the responses are handled.  Prior to the handle
changes, the loop connector could be completely purged of all matching
responses in flux_matchtag_free() via the purge connector method, but now
only the requeue queue in the generic handle code is purged when
matchtags are freed.  The loop connector is unique in that it always
had only one internal queue; other connectors would only purge their
internal requeue queues.  In any case, messages could always be in
flight and thus not subject to purge so this change is a step in the
right direction.

partially fixes #212