connectors/shmem: new connector for comms modules
This connector requires a shared zctx_t with the broker so it can
connect to the other end of inproc:// sockets.  Implement the
FLUX_OPT_ZEROMQ_CONTEXT setter in this connector, and defer socket
creation and connection until this option is set.  If the handle
is used before then, it is an EINVAL error.

Drop the specialized flux_get_zctx() getter and associated connector
methods.

Update the users:
- broker, when creating the broker end of the inproc:// socket
- modules/wrexec, which creates another zmq socket for wrexecd