libflux: add flux_opt_get/set() and connector methods
This provides a way to communicate abitrary info with a connector.
We will use this to pass the zmq context to the 'shmem' connector
so it can open the other end of an inproc socket already open
in the broker.

Option names are textual so an out of tree connector can define its
own without having to "register" it in handle.h.