libflux: flux_log_set_redirect takes callback
Instead of internally redirecting to src/common/libutil/log.c when
flux_log_set_redirect (true) is called, accept a callback function
instead of a boolean.

The broker registers such a callback on rank 0 so that log messages
can be sent to the configured destination (default stderr, changeable
with -L which accepts a file path, "stderr", "stdout", or "syslog".

Fixes #229.

Should make solving #67 easier since the callback preserves log
metadata that can be fed to syslog.