kvs: FASSERT if initial master commit failed
The kvs root sequence number is initialized to zero for an empty
root directory, then if there are any arguments passed to the kvs
main(), these are committed and the root sequence number is incremented.
Add an assertion that if there are arguments, the root sequence is > 0.

Debug for issue #238