connectors/local: fix pollfd super-readiness issue
This connector wrapped its socket fd in an edge triggered
epoll file descriptor, which it returned via its pollfd method.
Unfortunately, its pollevents method did not call epoll_wait()
to consume any epoll events, so this connector, once ready,
would continue to be ready forever and cause the reactor to spin.

Wrapping the socket fd is actually unnecessary; the fact that it
is level triggered is not a problem since the handle wraps both
the connector pollfd and the requeue pollfd in an edge triggered
epollfd.   Keep it simple and have the pollfd method return the
socket fd directly.