jsc: don't ignore the value passed to the watch callback
Partial fix for the races in Issue #234. JSC ignored the new
state value passed to the state-watch callback and instead
fetched the state from KVS on an state-chang event notification.

When the state changes rapidly, JSC gets a more updated
state. This is a straight out bug! But such a state drift issue
can affect other more complex notify-read patterns.

We are discussing a solution as part of Issue #206.