flux-exec: forward SIGTERM and SIGINT to all running processes
Modify signal handler in flux-exec to forward signals (SIGINT, SIGTERM
for now) to all running processes via the "cmb.exec.signal" rpc.

Note: there's a pretty clear race here if a signal is received before
processes are noted as running. The remote PID is required at this time
to signal a broker subprocess, so fixing the race completely will have to
wait a bit.