bindings/python:KVS and error checking repairs
fixed relative kvs directory acceses

kvs walk works from relative paths now

added a success_code to kvs calls

calls that pick up the spurious EBADF can now "safely" ignore it based
on the return value == the success code

Addressing error-checking issue from #253

This version checks the return from integer and pointer returning functions
and ensures that the value is a potential error before raising an exception on
errno.  To make the actual wrapper more extensible it is now a callable object
masquerading as a method instead of a closure.  This makes each a bit easier
to reason about.