bindings/python:test and kvs extensions python cmd bindings/python:fixing pre-refactor argument, adding walk test
bindings/python:Added specific-type gets to KVSDir

If you know what the target is, it's better to get it directly than have the
interface try and guess.

bindings/python:cleanup and formatting to address some of the more obvious pylint issues

bindings/python:updates to the message wrapper class

Handling of messages in echo was a little off.  Added a property to get/set
the message type and to get the type as a human-readable string, which helps
with debugging messages.

test/python:testing coverage, isolation and ergonomics improvements

This adds a coverage checking target, a large number of tests, and some general
ergonomics improvements to testing all over the place.  It also fixes some bugs
exposed by the new tests.  This version is sitting at 84% coverage of the
bindings, by number of statements.

altering check setup for python module detection

bindings/python:added a flux-python command

The new command invokes the python interpreter that flux was built with after
adding the flux installations exec prefix to the PYTHONPATH.  Nothing terribly
fancy, but it can be used to ensure that a script or command built on the flux
bindings will find them.

bindings/python: test fixups and re-factor

general cleanup