cmd/flux: initialze FLUX_CONNECTOR_PATH early
When configuration is being read from the KVS, and we are running
in-tree, it is necessary for FLUX_CONNECTOR_PATH to be set from the
in-tree config file before overriding the config from the KVS, or
flux_open() will not succeede in finding the local connector.

This allows "flux -T tmpdir" to once again work from outside a session.