test: add api disconnect test
Install a kvs watch then disconnect and see what happens to the
kvs reporting of the number of watchers.  Do this on a single node
using flux-kvs and flux comms-stats, and then use a specialized test
to do it across a session.  If successful this shows that disconnect
messages are generated for the same service on multiple nodes, and for
FLUX_NODEID_ANY.

Caveat: doesn't test FLUX_NODEID_UPSTREAM yet.