test: add helper script for travis builder
Add a helper script to build and install all dependencies for
continuous integration testing.

The script is meant to be general purpose and handle adding/updating
dependencies in an easy manner, but unfortunately lua deps have to
be special cased.

The script could potentially be used outside of travis-ci as well.