travis-ci: update to container based builder
travis-ci now has a container-based builder which is preferred over the
vm-based builders. The benefits of moving to this builder is faster build
start times, and additionally the ability to cache directories (e.g. for
built dependencies). The only drawback is that the use of sudo is no
longer allowed.

This update to .travis.yml activates the new container-based build by
setting

 sudo: false

in the .travis.yml file. We additionally use the src/test/travis-dep-builder.sh
script instead of manually building dependencies. This script handles the
management of the cache directories and install/build of depenedencies.

The end result should be much faster builds and checks.