cmd: exec: support stdin broadcast
Broadcast stdin to all running tasks.
There is a kludge here to wait until all tasks are started or failed
to process input. Probably a subprocess abstraction should be used in
flux-exec, with buffering until the processes are ready.

If flux-exec is sent to background, it will currently receive
SIGTTIN if it tries to read from the terminal. This will stop the
process, which can later be restarted. If starting the in background
and IO is not needed, it is suggested that the -n, --noinput
option be used to avoid this situation.