cleanup on exit
Added a utility API to clean up state on process exit.  Currently does files
and directories, but any kind of cleanup routine can be registered with the
API to be cleaned up when the process that registers the routine dies.  One
caveat, if that process execs, the cleanup will not get run.  You have been
warned.

fixes #71