libflux: add flux_msg_sendfd/recvfd
Add replacements for zfd_send(), zfd_recv() that use flux_msg_t
and handle non-blocking file descriptors properly:

  int flux_msg_sendfd (int fd, const flux_msg_t *msg,
                       struct flux_msg_iobuf *iobuf);
  flux_msg_t *flux_msg_recvfd (int fd, struct flux_msg_iobuf *iobuf);

On EOF, recvfd returns NULL with errno set to EPROTO.

If iobuf is non-NULL, and EWOULDBLOCK/EAGAIN is returned, you may
call the function again with identical arguments to continue reading/writing.
The 'iobuf' may be NULL if the file descriptor is in blocking mode.
If iobuf is NULL and an EWOULDBLOCK/EAGAIN is encountered, this is
internally converted to an EPROTO error, after which the channel
is in an undefined state.

iobuf should be initialized with flux_msg_iobuf_init() before use.

While EWOULDBLOCK/EAGAIN handling is in progress, iobuf contains
internally allocated storage.  If you need to dispose of this,
call flux_msg_iobuf_clean(), which is a no-op if no internal storage
is allocated.